fix(mobile): follow immich-31222's number formats through the Dart client
Giving every OpenAPI number property a format retypes the generated Dart client:
properties that pick up `integer` become `int` and ones that pick up
`format: double` become `double`, where both used to be `num`. Nothing in the
mobile source changed, so this only surfaces once the client is regenerated —
`dart analyze --fatal-infos` is the first thing that sees it.

Upstream's own half of immich-31222:

- `exif.converter.dart` (byte-identical to upstream) coerces latitude, longitude,
  fNumber and focalLength with `?.toDouble()`. Those are `double` now, so the
  call is a no-op and `noop_primitive_operations` is fatal here. Dropped.
- `RotateParameters.angle` is now `int`. `editor_test.dart` (also identical to
  upstream) builds one from `NormalizedTransform.rotation`, a double recovered
  from an affine matrix — so `.round()`, not `.toInt()`: acos/asin can land at
  89.999… for a quarter turn and truncation would send 89 to the server.

The fork's half, from this cycle's own `z.int()` on the Space counters and the
ratings facets:

- drop `.toInt()` from the space card's asset/member/new counts, the collection
  section's albumCount, and the space-person assetCount.
- retype the two `List<num> ratings` test helpers to `List<int>`.

`dart analyze --fatal-infos` is clean and `flutter test` is 3494 passing.

+++ mobile/lib/repositories/shared_space_api.repository.dart
@@ -140,7 +140,7 @@ class SharedSpaceApiRepository extends ApiRepository {
       name: dto.name,
       birthDate: dto.birthDate.orElse(null),
       spaceId: dto.spaceId,
-      numberOfAssets: dto.assetCount.toInt(),
+      numberOfAssets: dto.assetCount,
     );
   }
